fix circular import to allow importing models and metrics directly
