+++ eureka_ml_insights/configs/__init__.py
@@ -1,4 +1,3 @@
-from .aime import AIME_PIPELINE
 from .config import (
     AggregatorConfig,
     DataJoinConfig,
@@ -11,71 +10,7 @@ from .config import (
     PipelineConfig,
     PromptProcessingConfig,
 )
-from .dna import DNA_PIPELINE
-from .drop import Drop_Experiment_Pipeline
 from .experiment_config import ExperimentConfig, create_logdir
-from .flenqa import FlenQA_Experiment_Pipeline
-from .geometer import GEOMETER_PIPELINE
-from .gpqa import GPQA_Experiment_Pipeline
-from .ifeval import IFEval_PIPELINE
-from .image_understanding.object_detection import (
-    OBJECT_DETECTION_PAIRS_LOCAL_PIPELINE,
-    OBJECT_DETECTION_PAIRS_PIPELINE,
-    OBJECT_DETECTION_SINGLE_LOCAL_PIPELINE,
-    OBJECT_DETECTION_SINGLE_PIPELINE,
-)
-from .image_understanding.object_recognition import (
-    OBJECT_RECOGNITION_PAIRS_LOCAL_PIPELINE,
-    OBJECT_RECOGNITION_PAIRS_PIPELINE,
-    OBJECT_RECOGNITION_SINGLE_LOCAL_PIPELINE,
-    OBJECT_RECOGNITION_SINGLE_PIPELINE,
-)
-from .image_understanding.spatial_reasoning import (
-    SPATIAL_REASONING_PAIRS_LOCAL_PIPELINE,
-    SPATIAL_REASONING_PAIRS_PIPELINE,
-    SPATIAL_REASONING_SINGLE_LOCAL_PIPELINE,
-    SPATIAL_REASONING_SINGLE_PIPELINE,
-)
-from .image_understanding.visual_prompting import (
-    VISUAL_PROMPTING_PAIRS_LOCAL_PIPELINE,
-    VISUAL_PROMPTING_PAIRS_PIPELINE,
-    VISUAL_PROMPTING_SINGLE_LOCAL_PIPELINE,
-    VISUAL_PROMPTING_SINGLE_PIPELINE,
-)
-from .kitab import (
-    GPT35_KITAB_ONE_BOOK_CONSTRAINT_PIPELINE,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE_SELF_CONTEXT,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE_WITH_CONTEXT,
-    KITAB_TWO_BOOK_CONSTRAINT_PIPELINE,
-    KITAB_TWO_BOOK_CONSTRAINT_PIPELINE_WITH_CONTEXT,
-)
-from .mmmu import MMMU_BASELINE_PIPELINE
-from .nondeterminism import (
-    Geo_Nondeterminism,
-    IFEval_Nondeterminism,
-    Kitab_Nondeterminism,
-    MMMU_Nondeterminism,
-)
-from .toxigen import (
-    ToxiGen_Discriminative_PIPELINE,
-    ToxiGen_Generative_PIPELINE,
-)
-from .vision_language.maze import (
-    MAZE_PIPELINE,
-    MAZE_REPORTING_PIPELINE,
-    MAZE_TEXTONLY_PIPELINE,
-)
-from .vision_language.spatial_grid import (
-    SPATIAL_GRID_PIPELINE,
-    SPATIAL_GRID_REPORTING_PIPELINE,
-    SPATIAL_GRID_TEXTONLY_PIPELINE,
-)
-from .vision_language.spatial_map import (
-    SPATIAL_MAP_PIPELINE,
-    SPATIAL_MAP_REPORTING_PIPELINE,
-    SPATIAL_MAP_TEXTONLY_PIPELINE,
-)
 
 __all__ = [
     PipelineConfig,
@@ -89,50 +24,5 @@ __all__ = [
     DataSetConfig,
     EvalReportingConfig,
     ExperimentConfig,
-    OBJECT_DETECTION_PAIRS_PIPELINE,
-    OBJECT_DETECTION_SINGLE_PIPELINE,
-    OBJECT_DETECTION_PAIRS_LOCAL_PIPELINE,
-    OBJECT_DETECTION_SINGLE_LOCAL_PIPELINE,
-    OBJECT_RECOGNITION_PAIRS_PIPELINE,
-    OBJECT_RECOGNITION_SINGLE_PIPELINE,
-    OBJECT_RECOGNITION_PAIRS_LOCAL_PIPELINE,
-    OBJECT_RECOGNITION_SINGLE_LOCAL_PIPELINE,
-    SPATIAL_REASONING_PAIRS_PIPELINE,
-    SPATIAL_REASONING_SINGLE_PIPELINE,
-    SPATIAL_REASONING_PAIRS_LOCAL_PIPELINE,
-    SPATIAL_REASONING_SINGLE_LOCAL_PIPELINE,
-    VISUAL_PROMPTING_PAIRS_PIPELINE,
-    VISUAL_PROMPTING_SINGLE_PIPELINE,
-    VISUAL_PROMPTING_PAIRS_LOCAL_PIPELINE,
-    VISUAL_PROMPTING_SINGLE_LOCAL_PIPELINE,
-    SPATIAL_GRID_PIPELINE,
-    SPATIAL_GRID_TEXTONLY_PIPELINE,
-    SPATIAL_GRID_REPORTING_PIPELINE,
-    SPATIAL_MAP_PIPELINE,
-    SPATIAL_MAP_TEXTONLY_PIPELINE,
-    SPATIAL_MAP_REPORTING_PIPELINE,
-    MAZE_PIPELINE,
-    MAZE_TEXTONLY_PIPELINE,
-    MAZE_REPORTING_PIPELINE,
-    IFEval_PIPELINE,
-    FlenQA_Experiment_Pipeline,
-    GPQA_Experiment_Pipeline,
-    Drop_Experiment_Pipeline,
-    GEOMETER_PIPELINE,
-    MMMU_BASELINE_PIPELINE,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE_WITH_CONTEXT,
-    KITAB_ONE_BOOK_CONSTRAINT_PIPELINE_SELF_CONTEXT,
-    KITAB_TWO_BOOK_CONSTRAINT_PIPELINE,
-    KITAB_TWO_BOOK_CONSTRAINT_PIPELINE_WITH_CONTEXT,
-    GPT35_KITAB_ONE_BOOK_CONSTRAINT_PIPELINE,
-    DNA_PIPELINE,
-    ToxiGen_Discriminative_PIPELINE,
-    ToxiGen_Generative_PIPELINE,
-    Geo_Nondeterminism,
-    MMMU_Nondeterminism,
-    IFEval_Nondeterminism,
-    Kitab_Nondeterminism,
-    AIME_PIPELINE,
     create_logdir,
 ]
