+++ eureka_ml_insights/core/data_processing.py
@@ -1,11 +1,10 @@
-import base64
 import json
 import logging
 import os
 from hashlib import md5
 from typing import List, Optional
 
-import numpy as np
+from eureka_ml_insights.data_utils import NumpyEncoder
 
 from .pipeline import Component
 from .reserved_names import (
@@ -19,36 +18,6 @@ def compute_hash(val: str) -> str:
     Hashes the provided value using MD5.
     """
     return md5(val.encode("utf-8")).hexdigest()
-
-
-class NumpyEncoder(json.JSONEncoder):
-    """Special json encoder for numpy types"""
-
-    def default(self, obj):
-        if isinstance(
-            obj,
-            (
-                np.int_,
-                np.intc,
-                np.intp,
-                np.int8,
-                np.int16,
-                np.int32,
-                np.int64,
-                np.uint8,
-                np.uint16,
-                np.uint32,
-                np.uint64,
-            ),
-        ):
-            return int(obj)
-        elif isinstance(obj, (np.float_, np.float16, np.float32, np.float64)):
-            return float(obj)
-        elif isinstance(obj, (np.ndarray,)):
-            return obj.tolist()
-        elif isinstance(obj, bytes):
-            return base64.b64encode(obj).decode("ascii")
-        return json.JSONEncoder.default(self, obj)
 
 
 class DataProcessing(Component):
