+++ eureka_ml_insights/data_utils/__init__.py
@@ -11,8 +11,8 @@ from .data import (
     MMDataLoader,
     TXTWriter,
 )
+from .encoders import NumpyEncoder
 from .prompt_processing import JinjaPromptTemplate
-from .secret_key_utils import GetKey
 from .spatial_utils import (
     ExtractAnswerGrid,
     ExtractAnswerMaze,
@@ -67,11 +67,11 @@ __all__ = [
     RegexTransform,
     ASTEvalTransform,
     PrependStringTransform,
-    GetKey,
     ExtractAnswerGrid,
     ExtractAnswerSpatialMap,
     ExtractAnswerMaze,
     ShuffleColumnsTransform,
     ColumnMatchMapTransform,
     TokenCounterTransform,
+    NumpyEncoder,
 ]
