+++ eureka_ml_insights/user_configs/flenqa.py
@@ -17,7 +17,7 @@ from eureka_ml_insights.data_utils import (
 from eureka_ml_insights.data_utils.flenqa_utils import FlenQAOutputProcessor
 from eureka_ml_insights.metrics import CountAggregator, ExactMatch
 
-from .config import (
+from eureka_ml_insights.configs import (
     AggregatorConfig,
     DataProcessingConfig,
     DataSetConfig,
@@ -28,7 +28,7 @@ from .config import (
     PipelineConfig,
     PromptProcessingConfig,
 )
-from .experiment_config import ExperimentConfig
+from eureka_ml_insights.configs import ExperimentConfig
 
 
 class FlenQA_Experiment_Pipeline(ExperimentConfig):
