+++ eureka_ml_insights/user_configs/gpqa.py
@@ -15,7 +15,7 @@ from eureka_ml_insights.data_utils import (
 )
 from eureka_ml_insights.metrics import CountAggregator, ExactMatch
 
-from .config import (
+from eureka_ml_insights.configs import(
     AggregatorConfig,
     DataSetConfig,
     EvalReportingConfig,
@@ -25,7 +25,7 @@ from .config import (
     PipelineConfig,
     PromptProcessingConfig,
 )
-from .experiment_config import ExperimentConfig
+from eureka_ml_insights.configs import ExperimentConfig
 
 """This file contains user defined configuration classes for the geometric reasoning task on the GPQA dataset.
 """
