+++ eureka_ml_insights/user_configs/image_understanding/visual_prompting.py
@@ -3,31 +3,35 @@ import os
 from eureka_ml_insights.configs.experiment_config import ExperimentConfig
 from eureka_ml_insights.core import EvalReporting, Inference, PromptProcessing
 from eureka_ml_insights.data_utils import (
-    HFDataReader,    
+    HFDataReader,
     MMDataLoader,
     ColumnRename,
-    DataLoader,
     DataReader,
-    ExtractAnswerSpatialMap,
     PrependStringTransform,
     SequenceTransform,
 )
-from eureka_ml_insights.metrics import CaseInsensitiveMatch, CountAggregator
+from eureka_ml_insights.data_utils.spatial_utils import (
+    LowerCaseNoPunctuationConvertNumbers,
+)
+from eureka_ml_insights.metrics import (
+    CountAggregator,
+    ObjectRecognitionMetric,
+)
 
-from ..config import (
+from eureka_ml_insights.configs import (
     AggregatorConfig,
     DataSetConfig,
     EvalReportingConfig,
     InferenceConfig,
     MetricConfig,
-    ModelConfig,
     PipelineConfig,
     PromptProcessingConfig,
 )
+from .common import LOCAL_DATA_PIPELINE
 
-"""This file contains example user defined configuration classes for the spatial map task.
+"""This file contains example user defined configuration classes for the visual prompting task.
 In order to define a new configuration, a new class must be created that directly or indirectly
- inherits from UserDefinedConfig and the user_init method should be implemented.
+ inherits from ExperimentConfig and the configure_pipeline method should be implemented.
 You can inherit from one of the existing user defined classes below and override the necessary
 attributes to reduce the amount of code you need to write.
 
@@ -38,20 +42,22 @@ Pass the name of the class to the main.py script to run the pipeline.
 """
 
 
-class SPATIAL_MAP_PIPELINE(ExperimentConfig):
-    """This method is used to define an eval pipeline with inference and metric report components,
-    on the spatial map dataset."""
+class VISUAL_PROMPTING_PAIRS_PIPELINE(ExperimentConfig):
+    """
+    This defines an ExperimentConfig pipeline for the visual prompting dataset, pairs condition.
+    There is no model_config by default and the model config must be passed in via command lime.
+    """
 
-    def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
+    def configure_pipeline(self, model_config, resume_from=None):
         # Configure the data processing component.
         self.data_processing_comp = PromptProcessingConfig(
             component_type=PromptProcessing,
             data_reader_config=DataSetConfig(
                 HFDataReader,
                 {
-                    "path": "microsoft/VISION_LANGUAGE",
+                    "path": "microsoft/IMAGE_UNDERSTANDING",
                     "split": "val",
-                    "tasks": "spatial_map",
+                    "tasks": "visual_prompting_pairs",
                 },
             ),
             output_dir=os.path.join(self.log_dir, "data_processing_output"),
@@ -72,7 +78,6 @@ class SPATIAL_MAP_PIPELINE(ExperimentConfig):
         )
 
         # Configure the evaluation and reporting component.
-        # NOTE: This component uses model-specific answer extraction that is customized for GPT-4o, Claude, and Gemini models
         self.evalreporting_comp = EvalReportingConfig(
             component_type=EvalReporting,
             data_reader_config=DataSetConfig(
@@ -82,27 +87,15 @@ class SPATIAL_MAP_PIPELINE(ExperimentConfig):
                     "format": ".jsonl",
                     "transform": SequenceTransform(
                         [
-                            ColumnRename(name_mapping={"model_output": "model_output_raw"}),
-                            ExtractAnswerSpatialMap(
-                                answer_column_name="model_output_raw",
-                                extracted_answer_column_name="model_output",
-                                question_type_column_name="question_type",
-                                model_name=model_config.init_args['model_name'], # passing the model name for model-specific answer extraction
-                            ),
-                        ],
+                            LowerCaseNoPunctuationConvertNumbers(columns=["model_output"]),
+                        ]
                     ),
                 },
             ),
-            metric_config=MetricConfig(CaseInsensitiveMatch),
+            metric_config=MetricConfig(ObjectRecognitionMetric),
             aggregator_configs=[
-                AggregatorConfig(CountAggregator, {"column_names": ["CaseInsensitiveMatch_result"], "normalize": True}),
                 AggregatorConfig(
-                    CountAggregator,
-                    {
-                        "column_names": ["CaseInsensitiveMatch_result"],
-                        "group_by": "task",
-                        "normalize": True,
-                    },
+                    CountAggregator, {"column_names": ["ObjectRecognitionMetric_result"], "normalize": True}
                 ),
             ],
             output_dir=os.path.join(self.log_dir, "eval_report"),
@@ -112,22 +105,24 @@ class SPATIAL_MAP_PIPELINE(ExperimentConfig):
         return PipelineConfig([self.data_processing_comp, self.inference_comp, self.evalreporting_comp], self.log_dir)
 
 
-class SPATIAL_MAP_TEXTONLY_PIPELINE(SPATIAL_MAP_PIPELINE):
-    """This class extends SPATIAL_MAP_PIPELINE to use text only data."""
+class VISUAL_PROMPTING_SINGLE_PIPELINE(VISUAL_PROMPTING_PAIRS_PIPELINE):
+    """This class extends VISUAL_PROMPTING_PAIRS_PIPELINE to use the single object condition."""
 
-    def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
+    def configure_pipeline(self, model_config, resume_from=None):
         config = super().configure_pipeline(model_config, resume_from)
         self.data_processing_comp.data_reader_config.init_args["tasks"] = (
-            "spatial_map_text_only"
+            "visual_prompting_single"
         )
         return config
 
 
-class SPATIAL_MAP_REPORTING_PIPELINE(SPATIAL_MAP_PIPELINE):
-    """This method is used to define an eval pipeline with only a metric report component,
-    on the spatial map dataset."""
+class VISUAL_PROMPTING_PAIRS_LOCAL_PIPELINE(LOCAL_DATA_PIPELINE, VISUAL_PROMPTING_PAIRS_PIPELINE):
+    def configure_pipeline(self, model_config, resume_from=None):
+        local_path = "/home/neel/data/spatial_understanding"
+        return super().configure_pipeline(model_config, resume_from, local_path)
 
-    def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
-        super().configure_pipeline(model_config, resume_from)
-        self.evalreporting_comp.data_reader_config.init_args["path"] = resume_from
-        return PipelineConfig([self.evalreporting_comp], self.log_dir)
+
+class VISUAL_PROMPTING_SINGLE_LOCAL_PIPELINE(LOCAL_DATA_PIPELINE, VISUAL_PROMPTING_SINGLE_PIPELINE):
+    def configure_pipeline(self, model_config, resume_from=None):
+        local_path = "/home/neel/data/spatial_understanding"
+        return super().configure_pipeline(model_config, resume_from, local_path)
