+++ eureka_ml_insights/user_configs/specifications/image_understanding.txt
