+++ eureka_ml_insights/user_configs/vision_language/maze.py
@@ -8,12 +8,13 @@ from eureka_ml_insights.data_utils import (
     ColumnRename,
     DataLoader,
     DataReader,
-    ExtractAnswerGrid,
+    ExtractAnswerMaze,
     PrependStringTransform,
     SequenceTransform,
 )
 from eureka_ml_insights.metrics import CaseInsensitiveMatch, CountAggregator
-from ..config import (
+
+from eureka_ml_insights.configs import (
     AggregatorConfig,
     DataSetConfig,
     EvalReportingConfig,
@@ -24,7 +25,7 @@ from ..config import (
     PromptProcessingConfig,
 )
 
-"""This file contains example user defined configuration classes for the grid counting task.
+"""This file contains example user defined configuration classes for the maze task.
 In order to define a new configuration, a new class must be created that directly or indirectly
  inherits from UserDefinedConfig and the user_init method should be implemented.
 You can inherit from one of the existing user defined classes below and override the necessary
@@ -37,9 +38,9 @@ Pass the name of the class to the main.py script to run the pipeline.
 """
 
 
-class SPATIAL_GRID_PIPELINE(ExperimentConfig):
+class MAZE_PIPELINE(ExperimentConfig):
     """This method is used to define an eval pipeline with inference and metric report components,
-    on the grid counting dataset."""
+    on the spatial reasoning dataset."""
 
     def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
         # Configure the data processing component.
@@ -50,7 +51,7 @@ class SPATIAL_GRID_PIPELINE(ExperimentConfig):
                 {
                     "path": "microsoft/VISION_LANGUAGE",
                     "split": "val",
-                    "tasks": "spatial_grid",
+                    "tasks": "maze",
                 },
             ),
             output_dir=os.path.join(self.log_dir, "data_processing_output"),
@@ -81,11 +82,10 @@ class SPATIAL_GRID_PIPELINE(ExperimentConfig):
                     "transform": SequenceTransform(
                         [
                             ColumnRename(name_mapping={"model_output": "model_output_raw"}),
-                            ExtractAnswerGrid(
+                            ExtractAnswerMaze(
                                 answer_column_name="model_output_raw",
                                 extracted_answer_column_name="model_output",
                                 question_type_column_name="question_type",
-                                mode="animal",
                             ),
                         ],
                     ),
@@ -110,20 +110,20 @@ class SPATIAL_GRID_PIPELINE(ExperimentConfig):
         return PipelineConfig([self.data_processing_comp, self.inference_comp, self.evalreporting_comp], self.log_dir)
 
 
-class SPATIAL_GRID_TEXTONLY_PIPELINE(SPATIAL_GRID_PIPELINE):
-    """This class extends SPATIAL_GRID_PIPELINE to use text only data."""
+class MAZE_TEXTONLY_PIPELINE(MAZE_PIPELINE):
+    """This class extends MAZE_PIPELINE to use text only data."""
 
     def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
         config = super().configure_pipeline(model_config, resume_from)
         self.data_processing_comp.data_reader_config.init_args["tasks"] = (
-            "spatial_grid_text_only"
+            "maze_text_only"
         )
         return config
 
 
-class SPATIAL_GRID_REPORTING_PIPELINE(SPATIAL_GRID_PIPELINE):
+class MAZE_REPORTING_PIPELINE(MAZE_PIPELINE):
     """This method is used to define an eval pipeline with only a metric report component,
-    on the grid counting dataset."""
+    on the maze dataset."""
 
     def configure_pipeline(self, model_config: ModelConfig, resume_from: str = None) -> PipelineConfig:
         super().configure_pipeline(model_config, resume_from)
