+++ tests/pipeline_tests.py
@@ -12,7 +12,14 @@ logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(
 path = str(Path(Path(__file__).parent.absolute()).parent.absolute())  # noqa
 sys.path.insert(0, path)  # noqa
 
-from eureka_ml_insights.configs import (
+from eureka_ml_insights.configs import MetricConfig, ModelConfig
+from eureka_ml_insights.core import Pipeline
+from eureka_ml_insights.data_utils.transform import (
+    RunPythonTransform,
+    SamplerTransform,
+    SequenceTransform,
+)
+from eureka_ml_insights.user_configs import (
     AIME_PIPELINE,
     DNA_PIPELINE,
     GEOMETER_PIPELINE,
@@ -28,19 +35,11 @@ from eureka_ml_insights.configs import (
     SPATIAL_MAP_TEXTONLY_PIPELINE,
     SPATIAL_REASONING_SINGLE_PIPELINE,
     VISUAL_PROMPTING_SINGLE_PIPELINE,
-    GPQA_Experiment_Pipeline,
     Drop_Experiment_Pipeline,
+    GPQA_Experiment_Pipeline,
     IFEval_PIPELINE,
-    MetricConfig,
-    ModelConfig,
     ToxiGen_Discriminative_PIPELINE,
     ToxiGen_Generative_PIPELINE,
-)
-from eureka_ml_insights.core import Pipeline
-from eureka_ml_insights.data_utils.transform import (
-    RunPythonTransform,
-    SamplerTransform,
-    SequenceTransform,
 )
 from tests.test_utils import (
     DetectionTestModel,
@@ -262,6 +261,7 @@ class TEST_TOXIGEN_PIPELINE(ToxiGen_Discriminative_PIPELINE):
         }
         return config
 
+
 class TEST_TOXIGEN_GEN_PIPELINE(ToxiGen_Generative_PIPELINE):
     def configure_pipeline(self):
         config = super().configure_pipeline(model_config=ModelConfig(GenericTestModel, {}))
@@ -297,7 +297,8 @@ class TEST_GPQA_PIPELINE(GPQA_Experiment_Pipeline):
             "n_iter": N_ITER,
         }
         return config
-      
+
+
 class TEST_DROP_PIPELINE(Drop_Experiment_Pipeline):
     # Test config the Drop benchmark with TestModel and TestDataLoader
     def configure_pipeline(self):
@@ -462,13 +463,16 @@ class TOXIGEN_PipelineTest(PipelineTest, unittest.TestCase):
     def get_config(self):
         return TEST_TOXIGEN_PIPELINE().pipeline_config
 
+
 class TOXIGEN_GEN_PipelineTest(PipelineTest, unittest.TestCase):
     def get_config(self):
         return TEST_TOXIGEN_GEN_PIPELINE().pipeline_config
 
+
 class KITAB_ONE_BOOK_CONSTRAINT_PIPELINE_PipelineTest(PipelineTest, unittest.TestCase):
     def get_config(self):
         return TEST_KITAB_ONE_BOOK_CONSTRAINT_PIPELINE().pipeline_config
+
 
 @unittest.skipIf("skip_tests_with_missing_ds" in os.environ, "Missing public dataset. TODO: revert")
 class GPQA_PipelineTest(PipelineTest, unittest.TestCase):
